refactor: rename service locator
