+++ clean-architecture-news/lib/features/daily_news/presentation/pages/home/daily_news.dart
@@ -2,7 +2,7 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-import '../../../../../injector.dart';
+import '../../../../../service_locator.dart';
 import '../../../domain/entities/article.dart';
 import '../../bloc/article/remote/remote_article_bloc.dart';
 import '../../bloc/article/remote/remote_article_event.dart';
